DINOv2-Crack-Detection/requirements.txt
torch==1.20.0
torchvision==0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
einops>=0.4.0
pyyaml>=6.0
tensorboard>=2.13.0